#80879 - add event handler tests project
