+++ src/CaptainHook.sln
@@ -58,7 +58,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptainHook.TestsInfrastruc
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptainHook.Application.Infrastructure", "CaptainHook.Application.Infrastructure\CaptainHook.Application.Infrastructure.csproj", "{5D6010B6-5B3C-4A80-9FE6-8DCF2D7D96F6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CaptainHook.Domain.Tests", "Tests\CaptainHook.Domain.Tests\CaptainHook.Domain.Tests.csproj", "{41900278-11D4-48B6-98DE-410AD739CBE1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptainHook.Domain.Tests", "Tests\CaptainHook.Domain.Tests\CaptainHook.Domain.Tests.csproj", "{41900278-11D4-48B6-98DE-410AD739CBE1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CaptainHook.EventHandlerActor.Tests", "Tests\CaptainHook.EventHandlerActor.Tests\CaptainHook.EventHandlerActor.Tests.csproj", "{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -244,6 +246,14 @@ Global
 		{41900278-11D4-48B6-98DE-410AD739CBE1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{41900278-11D4-48B6-98DE-410AD739CBE1}.Release|x64.ActiveCfg = Release|Any CPU
 		{41900278-11D4-48B6-98DE-410AD739CBE1}.Release|x64.Build.0 = Release|Any CPU
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Debug|x64.ActiveCfg = Debug|x64
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Debug|x64.Build.0 = Debug|x64
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Release|x64.ActiveCfg = Release|x64
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -258,6 +268,7 @@ Global
 		{E6649154-892D-4CA2-A945-A864F6064881} = {24314874-C52E-4C8E-B7A5-D17EFD0C01DD}
 		{49D336DB-DDE6-43B8-BF41-48A27FD6992A} = {24314874-C52E-4C8E-B7A5-D17EFD0C01DD}
 		{41900278-11D4-48B6-98DE-410AD739CBE1} = {24314874-C52E-4C8E-B7A5-D17EFD0C01DD}
+		{8A8B1299-F2C7-4AC5-A153-76263A8A50FF} = {24314874-C52E-4C8E-B7A5-D17EFD0C01DD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {839C8B27-7C12-42F3-A1E3-85EB0F130282}
