+++ src/Tests/CaptainHook.Tests/Director/ConfigurationMergerTests.cs
@@ -7,7 +7,6 @@ using CaptainHook.Common.Configuration;
 using CaptainHook.Common.Configuration.KeyVault;
 using CaptainHook.DirectorService.Infrastructure;
 using CaptainHook.Domain.Entities;
-using CaptainHook.Tests.Builders;
 using CaptainHook.TestsInfrastructure.Builders;
 using Eshopworld.Tests.Core;
 using FluentAssertions;
