+++ src/Tests/CaptainHook.Tests/Director/DirectorServiceTests.cs
@@ -7,8 +7,8 @@ using CaptainHook.Common;
 using CaptainHook.Common.Configuration;
 using CaptainHook.DirectorService.Infrastructure.Interfaces;
 using CaptainHook.DirectorService.ReaderServiceManagement;
-using CaptainHook.Tests.Builders;
 using CaptainHook.Tests.Services;
+using CaptainHook.TestsInfrastructure.Builders;
 using Eshopworld.Core;
 using Eshopworld.Tests.Core;
 using FluentAssertions;
