+++ src/Tests/CaptainHook.Tests/Director/ReaderServiceManagement/ReaderServiceChangesDetectorTests.cs
@@ -2,7 +2,7 @@
 using CaptainHook.Common;
 using CaptainHook.Common.Configuration;
 using CaptainHook.DirectorService.ReaderServiceManagement;
-using CaptainHook.Tests.Builders;
+using CaptainHook.TestsInfrastructure.Builders;
 using Eshopworld.Tests.Core;
 using Xunit;
