+++ src/Tests/CaptainHook.TestsInfrastructure/Builders/SubscriberConfigurationBuilder.cs
@@ -4,11 +4,12 @@ using System.Net.Http;
 using CaptainHook.Common.Authentication;
 using CaptainHook.Common.Configuration;
 
-namespace CaptainHook.Tests.Builders
+namespace CaptainHook.TestsInfrastructure.Builders
 {
-    internal class WebhookConfigBuilder
+    public class SubscriberConfigurationBuilder
     {
         private string _type = "event1";
+        private string _subscriberName = "captain-hook";
         private string _uri = "https://blah.blah.eshopworld.com";
         private HttpMethod _httpMethod = HttpMethod.Post;
         private AuthenticationConfig _authenticationConfig = new BasicAuthenticationConfig
@@ -18,27 +19,34 @@ namespace CaptainHook.Tests.Builders
             Password = "password",
         };
         private List<WebhookRequestRule> _webhookRequestRules;
+        private WebhookConfig _callback;
+        private bool _asDlq;
 
-        public WebhookConfigBuilder WithType(string type)
+        public SubscriberConfigurationBuilder WithType(string type)
         {
             _type = type;
             return this;
         }
 
-        public WebhookConfigBuilder WithUri(string uri)
+        public SubscriberConfigurationBuilder WithSubscriberName(string subscriberName)
+        {
+            _subscriberName = subscriberName;
+            return this;
+        }
+
+        public SubscriberConfigurationBuilder WithUri(string uri)
         {
             _uri = uri;
             return this;
         }
 
-        public WebhookConfigBuilder WithHttpVerb(string httpVerb)
+        public SubscriberConfigurationBuilder AsDLQ (bool asDlq = true)
         {
-            _httpMethod = new HttpMethod(httpVerb);
-
+            _asDlq = asDlq;
             return this;
         }
 
-        public WebhookConfigBuilder WithOidcAuthentication()
+        public SubscriberConfigurationBuilder WithOidcAuthentication()
         {
             _authenticationConfig = new OidcAuthenticationConfig
             {
@@ -52,7 +60,7 @@ namespace CaptainHook.Tests.Builders
             return this;
         }
 
-        public WebhookConfigBuilder WithBasicAuthentication()
+        public SubscriberConfigurationBuilder WithBasicAuthentication()
         {
             _authenticationConfig = new BasicAuthenticationConfig
             {
@@ -64,28 +72,33 @@ namespace CaptainHook.Tests.Builders
             return this;
         }
 
-        public WebhookConfigBuilder WithNoAuthentication()
+        // TODO: remove or use version of this method which accepts SubscriberConfigurationBuilder
+        public SubscriberConfigurationBuilder WithCallback(string uri = "https://callback.eshopworld.com")
         {
-            _authenticationConfig = new AuthenticationConfig
+            _callback = new WebhookConfig
             {
-                Type = AuthenticationType.None
+                Name = "callback",
+                HttpMethod = HttpMethod.Post,
+                Uri = uri,
+                EventType = "event1",
+                AuthenticationConfig = new AuthenticationConfig
+                {
+                    Type = AuthenticationType.None
+                },
             };
 
             return this;
         }
 
-        public WebhookConfigBuilder SetWebhookRequestRule(Action<WebhookRequestRuleBuilder> ruleBuilder)
-        {            
-            var builder = new WebhookRequestRuleBuilder();
-            ruleBuilder(builder);
-            _webhookRequestRules = new List<WebhookRequestRule>
-            {
-                builder.Create()
-            };
+        public SubscriberConfigurationBuilder WithCallback(Action<SubscriberConfigurationBuilder> callbackBuilder)
+        {
+            var builder = new SubscriberConfigurationBuilder();
+            callbackBuilder(builder);
+            _callback = builder.Create();
             return this;
         }
 
-        public WebhookConfigBuilder AddWebhookRequestRule(Action<WebhookRequestRuleBuilder> ruleBuilder)
+        public SubscriberConfigurationBuilder AddWebhookRequestRule(Action<WebhookRequestRuleBuilder> ruleBuilder)
         {
             if (_webhookRequestRules == null)
             {
@@ -98,19 +111,23 @@ namespace CaptainHook.Tests.Builders
             return this;
         }
 
-        public WebhookConfig Create()
+        public SubscriberConfiguration Create()
         {
-            var webhookConfig = new WebhookConfig
+            var subscriber = new SubscriberConfiguration
             {
                 Name = _type,
                 EventType = _type,
+                SubscriberName = _subscriberName,
                 HttpMethod = _httpMethod,
                 Uri = _uri,
                 AuthenticationConfig = _authenticationConfig,
+                Callback = _callback,
                 WebhookRequestRules = _webhookRequestRules,
+                DLQMode = _asDlq ? SubscriberDlqMode.WebHookMode: (SubscriberDlqMode?) null
             };
 
-            return webhookConfig;
+            return subscriber;
         }
+
     }
 }
